test lib/create-description-box.js
