+++ lib/transforms/search-results-to-template-data.js
@@ -10,9 +10,7 @@ const pillboxImage = require('../helpers/search-results-to-template-data/pillbox
 const getTitlePage = require('../helpers/search-results-to-template-data/title-page.js');
 const paramify = require('../helpers/paramify');
 const querify = require('../helpers/querify');
-const categorysDescription = require('../../description-boxes/category.json');
-const gallerysDescription = require('../../description-boxes/gallery.json');
-const museumsDescription = require('../../description-boxes/museum.json');
+const createDescriptionBox = require('../create-description-box.js');
 
 module.exports = (queryParams, results, config) => {
   config = config || {};
@@ -207,44 +205,4 @@ function createClearStickyFilterLink (queryParams, rootUrl, filter) {
   params['page[type]'] = queryParams.pageType;
   const type = queryParams.type !== 'all' ? '/' + queryParams.type : '';
   return `${rootUrl}/search${type}${paramify(params)}${querify(params)}`;
-}
-
-function createDescriptionBox (selectedFilters) {
-  selectedFilters = selectedFilters || {};
-
-  function getDescription (filterKey, jsonDescription) {
-    const filter = Object.keys(selectedFilters[filterKey] || [])[0];
-    return jsonDescription[filter];
-  }
-
-  function filterNotFoundMessage (filterType, filterKey) {
-    const shouldMatch = Object.keys(selectedFilters[filterKey] || [])[0];
-    console.log(filterType + ' didn\'t match, ensure the exact ' +
-      filterType.toLowerCase() + ' title was used');
-    console.log(filterKey + ' title should match: ', shouldMatch);
-  }
-
-  if (selectedFilters.categories) {
-    const categoryDescription = getDescription('categories', categorysDescription);
-    if (categoryDescription) {
-      return {category: categoryDescription};
-    }
-    filterNotFoundMessage('Category', 'categories');
-  }
-
-  if (selectedFilters.gallery) {
-    const galleryDescription = getDescription('gallery', gallerysDescription);
-    if (galleryDescription) {
-      return {gallery: galleryDescription};
-    }
-    filterNotFoundMessage('Gallery', 'gallery');
-  }
-
-  if (selectedFilters.museum) {
-    const museumDescription = getDescription('museum', museumsDescription);
-    if (museumDescription) {
-      return {museum: museumDescription};
-    }
-    filterNotFoundMessage('Museum', 'museum');
-  }
 }
